Adds section navigation.
Section navigation is now automatically generated based on the section
field of the Front Matter metadata. I have to tweak it slightly after
the fact so that it appears properly in the site, but with some more
work the transformations could be described in the platform site
metadata.

By default, pages are displayed in the nav as links displaying the
page title. This string can be overridden using the navTitle field.

A page can override its nav entirely by setting the navigation field.